re-organize helper functions for publicationsData
