--- packages/app-root/src/app/about/publications/projectAvatarsMap.mock.json
@@ -1,29 +1,0 @@
-{
-  "1": {
-    "id": "1",
-    "display_name": "Galaxy Zoo",
-    "updated_at": "2019-12-05T09:28:39.807Z",
-    "description": "Help us discover the secrets of galaxy evolution by classifying distant galaxies.",
-    "slug": "zooniverse/galaxy-zoo",
-    "redirect": "http://galaxyzoo.org",
-    "avatar_src": "panoptes-uploads.zooniverse.org/production/project_avatar/8e945c4f-4d91-46b8-875e-cdfe28a0c66e.jpeg"
-  },
-  "2": {
-    "id": "2",
-    "display_name": "Planet Hunters",
-    "updated_at": "2019-12-05T09:26:19.152Z",
-    "description": "Find planets around stars. Lightcurve changes from the Kepler spacecraft can indicate transiting planets. ",
-    "slug": "zooniverse/planet-hunters",
-    "redirect": "http://planethunters.org/#/",
-    "avatar_src": "panoptes-uploads.zooniverse.org/production/project_avatar/912d988e-19c3-4abf-b632-312453042d8a.jpeg"
-  },
-  "3": {
-    "id": "3",
-    "display_name": "Snapshot Serengeti",
-    "updated_at": "2019-12-05T09:27:20.052Z",
-    "description": "Help us explore the secret life of the Serengeti! Our cameras track the movement of thousands upon thousands of animals migrating with and following the wildebeest migration.",
-    "slug": "zooniverse/snapshot-serengeti",
-    "redirect": "",
-    "avatar_src": "panoptes-uploads.zooniverse.org/project_avatar/c05aa83a-cc54-455a-bde6-a42c4a84a702.jpeg"
-  }
-}
