--- packages/app-root/src/app/about/publications/uniqueProjectIds.mock.json
@@ -1,1 +1,0 @@
-["1", "2", "3"]
